[REFACTOR] introduce domain entities and ports
